server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Supply Daddy — Decentralized logistics backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn[standard]>=0.30.0",
    "web3>=7.0.0",
    "pdfplumber>=0.11.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.0.0",
]
